+++ test/Spec.hs
@@ -1,6 +1,8 @@
 module Main (main) where
 
 import qualified IR.Domain.ErrorSpec
+import qualified IR.Domain.Module.DockerSpec
+import qualified IR.Domain.Module.PostgreSQLSpec
 import qualified IR.Domain.ModuleSpec
 import qualified IR.Domain.NetworkSpec
 import qualified IR.Domain.PackageSpec
@@ -17,6 +19,8 @@ main =
         testGroup
             "IR"
             [ IR.Domain.ErrorSpec.tests
+            , IR.Domain.Module.DockerSpec.tests
+            , IR.Domain.Module.PostgreSQLSpec.tests
             , IR.Domain.ModuleSpec.tests
             , IR.Domain.NetworkSpec.tests
             , IR.Domain.PackageSpec.tests